requirements.txt
flask
tensorflow
keras
numpy
pillow
python-dotenv